Android2 homework7

+++ app/src/main/java/ru/geekbrains/myweatherapp/weatherData/Wind.java
@@ -8,12 +8,12 @@ public class Wind {
 
     @SerializedName("speed")
     @Expose
-    private Integer speed;
+    private double speed;
     @SerializedName("deg")
     @Expose
     private Integer deg;
 
-    public Integer getSpeed() {
+    public double getSpeed() {
         return speed;
     }
